backend/tests/test_api_endpoints.py
"""Test API endpoint logic - unit tests (no database required)."""

import pytest
from unittest.mock import AsyncMock, patch


class TestAPIEndpointLogic:
    """Test API endpoint query logic (unit tests)."""
    
    def test_get_games_count_logic(self):
        """Test /games count query structure."""
        # Verify the query would work (syntax check)
        query = "SELECT COUNT(*) FROM game"
        assert "COUNT" in query
        assert "game" in query
    
    def test_get_competitions_with_filters_logic(self):
        """Test /competitions granular fields query structure."""
        query = """
            SELECT discipline, series, "group", pool, season, year, organization
            FROM competition
            WHERE discipline IS NOT NULL
        """
        assert "discipline" in query
        assert "series" in query
        assert "group" in query
        assert "pool" in query
    
    def test_get_competitions_filters_logic(self):
        """Test /competitions/filters query structure."""
        queries = [
            "SELECT DISTINCT discipline FROM competition",
            "SELECT DISTINCT season FROM competition",
            "SELECT DISTINCT year FROM competition",
            "SELECT DISTINCT series FROM competition",
            'SELECT DISTINCT "group" FROM competition',
            "SELECT DISTINCT pool FROM competition",
        ]
        
        for query in queries:
            assert "SELECT DISTINCT" in query
            assert "competition" in query
    
    def test_get_clubs_logic(self):
        """Test /clubs query structure."""
        query = "SELECT id, name, is_active FROM club WHERE is_active = true"
        assert "club" in query
        assert "is_active" in query


class TestAPIEndpointFilterLogic:
    """Test API filter query logic (unit tests)."""
    
    def test_filter_by_discipline_and_year_logic(self):
        """Test filtering by discipline and year - query structure."""
        query = """
            SELECT id, discipline, year
            FROM competition
            WHERE discipline IS NOT NULL AND year IS NOT NULL
        """
        assert "discipline" in query
        assert "year" in query
        assert "WHERE" in query
    
    def test_filter_by_series_and_group_logic(self):
        """Test filtering by series and group - query structure."""
        query = """
            SELECT id, series, "group"
            FROM competition
            WHERE series IS NOT NULL OR "group" IS NOT NULL
        """
        assert "series" in query
        assert "group" in query
        assert "WHERE" in query
