backend/pyproject.toml
[project]
name = "pilota-backend"
version = "0.1.0"
description = "Basque Pelota scraping and API backend"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.14.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.112.0",
    "httpx>=0.27.0",
    "pre-commit>=4.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.30.0",
    "aiohttp>=3.9.0",
    "lxml>=5.0.0",
    "requests>=2.31.0",
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
